requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-groq>=0.0.1
langchain-anthropic>=0.0.1
faiss-cpu>=1.7.4
pypdf2>=3.0.1
python-docx>=0.8.11
markdown>=3.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
slack-bolt>=1.18.0
numpy>=1.26.0
pandas>=2.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0 